Refactor: 총 수익률 반환 기능 클래스 이전
- 총 수익률을 반환해주는 기능은 구매금액을 관리하는 BuyCash 클래스에서 관리하는게 연관성이 있다고 판단함.
- 기능과 테스트 코드를 이전함.
- 사용을 안하게된 LottoService클래스를 제거함.
