--- src/main/java/lotto/Constant.java
@@ -1,5 +1,0 @@
-package lotto;
-
-public class Constant {
-    public static final int BUY_UNIT=1000;
-}
