+++ src/test/java/lotto/BuyCashTest.java
@@ -4,6 +4,7 @@ import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
 
 import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 
@@ -26,5 +27,20 @@ public class BuyCashTest {
         BuyCash buyCash = new BuyCash(cash);
 
         assertThat(buyCash.getLottoAmount()).isEqualTo(expected);
+    }
+
+    @DisplayName("총 수익과 구매금액을 통해 수익률을 반환해주는 기능.")
+    @Test
+    void 수익률_반환() {
+        //given
+        long totalProfit = 5000;
+        BuyCash buyCash = new BuyCash(8000);
+        double expected = 62.5;
+
+        //when
+        double result = buyCash.getTotalProfitRate(totalProfit);
+
+        //then
+        assertThat(result).isEqualTo(expected);
     }
 }
